requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
motor==3.3.2
pymongo==4.6.0
httpx==0.25.2
jinja2==3.1.2
python-dotenv==1.0.0
